refactor: Move layout code from AppContainer to LayoutObserver, clean up
